requirements.txt
# --------- pytorch --------- #
torch>=2.9.1
torchvision>=0.24
lightning>=2.5.5

# --------- hydra --------- #
hydra-core
hydra-colorlog
hydra-optuna-sweeper
omegaconf

# --------- others --------- #
rootutils       # standardizing the project root setup
rich            # beautiful text formatting in terminal
pytest          # tests
tensorboard     # logging training metrics
mlflow          # logging using MLFlow
loguru          # simple logging
